Clang: enable -Wtautological-constant-out-of-range-compare
This warning was implemented in Clang a couple of weeks ago,
and now our code base is clean enough to enable it.

BUG=151927


Review URL: https://chromiumcodereview.appspot.com/11366060

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@165793 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 21c35d7280ba62564ad4c2087438501f01ee3420

+++ chrome/browser/chromeos/web_socket_proxy.cc
@@ -1310,7 +1310,7 @@ bool Conn::EmitFrame(
   uint8 header[10];
   int header_size = 2;
   DCHECK(chan);
-  DCHECK(opcode >= 0 && opcode < 16);
+  DCHECK(opcode >= 0 && static_cast<int>(opcode) < 16);
   header[0] = opcode | 0x80;  // FIN bit set.
   if (size < 126) {
     header[1] = size;
